requirements.txt
langchain
langchain-openai
langchain-community
slack_sdk
python-dotenv
scipy
tiktoken
pandas